+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 72.116.0 - 2022-06-23
+* [#1484](https://github.com/stripe/stripe-go/pull/1484) API Updates
+  * Add support for `CaptureMethod` on `PaymentIntentConfirmParams` and `PaymentIntentParams`
+* [#1481](https://github.com/stripe/stripe-go/pull/1481) API Updates
+  * Add support for `PromptPayPayments` on `AccountCapabilitiesParams` and `AccountCapabilities`
+  * Add support for `PromptPay` on `ChargePaymentMethodDetails`, `PaymentIntentConfirmPaymentMethodDataParams`, `PaymentIntentConfirmPaymentMethodOptionsParams`, `PaymentIntentPaymentMethodDataParams`, `PaymentIntentPaymentMethodOptionsParams`, `PaymentIntentPaymentMethodOptions`, `PaymentMethodParams`, `PaymentMethod`, `SetupIntentConfirmPaymentMethodDataParams`, and `SetupIntentPaymentMethodDataParams`
+  * Add support for `SubtotalExcludingTax` on `CreditNote` and `Invoice`
+  * Add support for `AmountExcludingTax` and `UnitAmountExcludingTax` on `CreditNoteLineItem` and `InvoiceLineItem`
+  * Add support for `RenderingOptions` on `InvoiceParams`
+  * Add support for `TotalExcludingTax` on `Invoice`
+  * Add support for new value `promptpay` on enums `InvoicePaymentSettingsPaymentMethodTypes` and `SubscriptionPaymentSettingsPaymentMethodTypes`
+  * Add support for `AutomaticPaymentMethods` on `OrderPaymentSettings`
+  * Add support for `PromptPayDisplayQRCode` on `PaymentIntentNextAction`
+  * Add support for new value `promptpay` on enum `PaymentMethodType`
+  
+* [#1482](https://github.com/stripe/stripe-go/pull/1482) Use the generated API version
+
 ## 72.115.0 - 2022-06-17
 * [#1477](https://github.com/stripe/stripe-go/pull/1477) API Updates
   * Add support for `FundCashBalance` test helper method on resource `Customer`